requirements.txt
supabase
flask
pyngrok
flask
twilio
python-dotenv
inquirer
black
<<<<<<< HEAD
openai
=======
>>>>>>> master
